Fix issue 9: Incorrect polling locations are being returned
Prior to this change, the code was assuming that the polling location
and ward/precinct indices were the same. This does not seem to be the
case, as while some addresses would return the correct polling place,
others would not. Clearly, the data is not as consistent as we assumed.

In order to fix the issue, we're now searching the polling places, and
matching them directly against the precinct and ward returned by the
precinct and ward data.